+++ lib/src/models/credit_card_config.dart
@@ -2,6 +2,8 @@
 class CreditCardConfig {
   ///  An option to save (tokenize) the card after a successful payment.
   bool saveCard;
+
+  /// An option to enable the manual auth and capture
   bool manual;
 
   CreditCardConfig({required this.saveCard, required this.manual});